+++ projects/ngrx-auto-entity/src/lib/operators.ts
@@ -75,17 +75,20 @@ export const handleError = <TModel, TErrorAction>(
     const serviceName = `${pascalCase(error.info.modelName)}Service`;
     console.error(
       `[NGRX-AE] ! NgRxAutoEntityService Error: Unable to locate load method in the ${serviceName}`,
+      '\nReason: ',
       error.err
     );
   } else if (error.info && error.message) {
     const serviceName = `${pascalCase(error.info.modelName)}Service`;
     console.error(
       `[NGRX-AE] ! NgRxAutoEntityService Error: Unable to invoke required operations on the ${serviceName}`,
+      '\nReason: ',
       error.message
     );
   } else if (error.message) {
     console.error(
       `[NGRX-AE] ! NgRxAutoEntityService Error: Unable to invoke required operations on entity service`,
+      '\nReason: ',
       error.message
     );
   } else {
@@ -104,11 +107,11 @@ export class EntityOperators {
   load<TModel>() {
     return (source: Observable<Load<TModel>>) =>
       source.pipe(
-        mergeMap(({ info, keys, criteria }) => {
+        mergeMap(({ info, keys, criteria, correlationId }) => {
           return this.entityService.load(info, keys, criteria).pipe(
-            map((ref: IEntityRef<TModel>) => new LoadSuccess<TModel>(ref.info.modelType, ref.entity)),
+            map((ref: IEntityRef<TModel>) => new LoadSuccess<TModel>(ref.info.modelType, ref.entity, correlationId)),
             catchError((error: IEntityError<TModel>) =>
-              handleError(error, new LoadFailure<TModel>(error.info.modelType, error.err))
+              handleError(error, new LoadFailure<TModel>(error.info.modelType, error.err, correlationId))
             )
           );
         })
@@ -118,11 +121,13 @@ export class EntityOperators {
   loadAll<TModel>() {
     return (source: Observable<LoadAll<TModel>>) =>
       source.pipe(
-        mergeMap(({ info, criteria }) => {
+        mergeMap(({ info, criteria, correlationId }) => {
           return this.entityService.loadAll(info, criteria).pipe(
-            map((ref: IEntityRef<TModel[]>) => new LoadAllSuccess<TModel>(ref.info.modelType, ref.entity)),
+            map(
+              (ref: IEntityRef<TModel[]>) => new LoadAllSuccess<TModel>(ref.info.modelType, ref.entity, correlationId)
+            ),
             catchError((error: IEntityError<TModel>) =>
-              handleError(error, new LoadAllFailure<TModel>(error.info.modelType, error.err))
+              handleError(error, new LoadAllFailure<TModel>(error.info.modelType, error.err, correlationId))
             )
           );
         })
@@ -132,11 +137,11 @@ export class EntityOperators {
   loadMany<TModel>() {
     return (source: Observable<LoadMany<TModel>>) =>
       source.pipe(
-        mergeMap(({ info, criteria }) => {
+        mergeMap(({ info, criteria, correlationId }) => {
           return this.entityService.loadMany(info, criteria).pipe(
-            map((ref: IEntityRef<TModel[]>) => new LoadManySuccess<TModel>(ref.info.modelType, ref.entity)),
+            map((ref: IEntityRef<TModel[]>) => new LoadManySuccess<TModel>(ref.info.modelType, ref.entity, correlationId)),
             catchError((error: IEntityError<TModel>) =>
-              handleError(error, new LoadManyFailure<TModel>(error.info.modelType, error.err))
+              handleError(error, new LoadManyFailure<TModel>(error.info.modelType, error.err, correlationId))
             )
           );
         })
@@ -146,13 +151,13 @@ export class EntityOperators {
   loadPage<TModel>() {
     return (source: Observable<LoadPage<TModel>>) =>
       source.pipe(
-        mergeMap(({ info, page, criteria }) => {
+        mergeMap(({ info, page, criteria, correlationId }) => {
           return this.entityService.loadPage(info, page, criteria).pipe(
             map(
-              (ref: IEntityPageRef<TModel>) => new LoadPageSuccess<TModel>(ref.info.modelType, ref.entity, ref.pageInfo)
+              (ref: IEntityPageRef<TModel>) => new LoadPageSuccess<TModel>(ref.info.modelType, ref.entity, ref.pageInfo, correlationId)
             ),
             catchError((error: IEntityError<TModel>) =>
-              handleError(error, new LoadPageFailure<TModel>(error.info.modelType, error.err))
+              handleError(error, new LoadPageFailure<TModel>(error.info.modelType, error.err, correlationId))
             )
           );
         })
@@ -162,14 +167,14 @@ export class EntityOperators {
   loadRange<TModel>() {
     return (source: Observable<LoadRange<TModel>>) =>
       source.pipe(
-        mergeMap(({ info, range, criteria }) => {
+        mergeMap(({ info, range, criteria, correlationId }) => {
           return this.entityService.loadRange(info, range, criteria).pipe(
             map(
               (ref: IEntityRangeRef<TModel>) =>
-                new LoadRangeSuccess<TModel>(ref.info.modelType, ref.entity, ref.rangeInfo)
+                new LoadRangeSuccess<TModel>(ref.info.modelType, ref.entity, ref.rangeInfo, correlationId)
             ),
             catchError((error: IEntityError<TModel>) =>
-              handleError(error, new LoadRangeFailure<TModel>(error.info.modelType, error.err))
+              handleError(error, new LoadRangeFailure<TModel>(error.info.modelType, error.err, correlationId))
             )
           );
         })
@@ -179,11 +184,11 @@ export class EntityOperators {
   create<TModel>() {
     return (source: Observable<Create<TModel>>) =>
       source.pipe(
-        mergeMap(({ info, entity, criteria }) => {
+        mergeMap(({ info, entity, criteria, correlationId }) => {
           return this.entityService.create<TModel>(info, entity, criteria).pipe(
-            map((ref: IEntityRef<TModel>) => new CreateSuccess<TModel>(ref.info.modelType, ref.entity)),
+            map((ref: IEntityRef<TModel>) => new CreateSuccess<TModel>(ref.info.modelType, ref.entity, correlationId)),
             catchError((error: IEntityError<TModel>) =>
-              handleError(error, new CreateFailure<TModel>(error.info.modelType, error.err))
+              handleError(error, new CreateFailure<TModel>(error.info.modelType, error.err, correlationId))
             )
           );
         })
@@ -193,11 +198,11 @@ export class EntityOperators {
   createMany<TModel>() {
     return (source: Observable<CreateMany<TModel>>) =>
       source.pipe(
-        mergeMap(({ info, entities, criteria }) => {
+        mergeMap(({ info, entities, criteria, correlationId }) => {
           return this.entityService.createMany<TModel>(info, entities, criteria).pipe(
-            map((ref: IEntityRef<TModel[]>) => new CreateManySuccess<TModel>(ref.info.modelType, ref.entity)),
+            map((ref: IEntityRef<TModel[]>) => new CreateManySuccess<TModel>(ref.info.modelType, ref.entity, correlationId)),
             catchError((error: IEntityError<TModel>) =>
-              handleError(error, new CreateManyFailure<TModel>(error.info.modelType, error.err))
+              handleError(error, new CreateManyFailure<TModel>(error.info.modelType, error.err, correlationId))
             )
           );
         })
@@ -207,11 +212,11 @@ export class EntityOperators {
   update<TModel>() {
     return (source: Observable<Update<TModel>>) =>
       source.pipe(
-        mergeMap(({ info, entity, criteria }) => {
+        mergeMap(({ info, entity, criteria, correlationId }) => {
           return this.entityService.update<TModel>(info, entity, criteria).pipe(
-            map((ref: IEntityRef<TModel>) => new UpdateSuccess<TModel>(ref.info.modelType, ref.entity)),
+            map((ref: IEntityRef<TModel>) => new UpdateSuccess<TModel>(ref.info.modelType, ref.entity, correlationId)),
             catchError((error: IEntityError<TModel>) =>
-              handleError(error, new UpdateFailure<TModel>(error.info.modelType, error.err))
+              handleError(error, new UpdateFailure<TModel>(error.info.modelType, error.err, correlationId))
             )
           );
         })
@@ -221,11 +226,11 @@ export class EntityOperators {
   updateMany<TModel>() {
     return (source: Observable<UpdateMany<TModel>>) =>
       source.pipe(
-        mergeMap(({ info, entities, criteria }) => {
+        mergeMap(({ info, entities, criteria, correlationId }) => {
           return this.entityService.updateMany<TModel>(info, entities, criteria).pipe(
-            map((ref: IEntityRef<TModel[]>) => new UpdateManySuccess<TModel>(ref.info.modelType, ref.entity)),
+            map((ref: IEntityRef<TModel[]>) => new UpdateManySuccess<TModel>(ref.info.modelType, ref.entity, correlationId)),
             catchError((error: IEntityError<TModel>) =>
-              handleError(error, new UpdateManyFailure<TModel>(error.info.modelType, error.err))
+              handleError(error, new UpdateManyFailure<TModel>(error.info.modelType, error.err, correlationId))
             )
           );
         })
@@ -235,11 +240,11 @@ export class EntityOperators {
   replace<TModel>() {
     return (source: Observable<Replace<TModel>>) =>
       source.pipe(
-        mergeMap(({ info, entity, criteria }) => {
+        mergeMap(({ info, entity, criteria, correlationId }) => {
           return this.entityService.replace<TModel>(info, entity, criteria).pipe(
-            map((ref: IEntityRef<TModel>) => new ReplaceSuccess<TModel>(ref.info.modelType, ref.entity)),
+            map((ref: IEntityRef<TModel>) => new ReplaceSuccess<TModel>(ref.info.modelType, ref.entity, correlationId)),
             catchError((error: IEntityError<TModel>) =>
-              handleError(error, new ReplaceFailure<TModel>(error.info.modelType, error.err))
+              handleError(error, new ReplaceFailure<TModel>(error.info.modelType, error.err, correlationId))
             )
           );
         })
@@ -249,11 +254,11 @@ export class EntityOperators {
   replaceMany<TModel>() {
     return (source: Observable<ReplaceMany<TModel>>) =>
       source.pipe(
-        mergeMap(({ info, entities, criteria }) => {
+        mergeMap(({ info, entities, criteria, correlationId }) => {
           return this.entityService.replaceMany<TModel>(info, entities, criteria).pipe(
-            map((ref: IEntityRef<TModel[]>) => new ReplaceManySuccess<TModel>(ref.info.modelType, ref.entity)),
+            map((ref: IEntityRef<TModel[]>) => new ReplaceManySuccess<TModel>(ref.info.modelType, ref.entity, correlationId)),
             catchError((error: IEntityError<TModel>) =>
-              handleError(error, new ReplaceManyFailure<TModel>(error.info.modelType, error.err))
+              handleError(error, new ReplaceManyFailure<TModel>(error.info.modelType, error.err, correlationId))
             )
           );
         })
@@ -263,11 +268,11 @@ export class EntityOperators {
   delete<TModel>() {
     return (source: Observable<Delete<TModel>>) =>
       source.pipe(
-        mergeMap(({ info, entity, criteria }) => {
+        mergeMap(({ info, entity, criteria, correlationId }) => {
           return this.entityService.delete(info, entity, criteria).pipe(
-            map((ref: IEntityRef<TModel>) => new DeleteSuccess<TModel>(ref.info.modelType, ref.entity)),
+            map((ref: IEntityRef<TModel>) => new DeleteSuccess<TModel>(ref.info.modelType, ref.entity, correlationId)),
             catchError((error: IEntityError<TModel>) =>
-              handleError(error, new DeleteFailure<TModel>(error.info.modelType, error.err))
+              handleError(error, new DeleteFailure<TModel>(error.info.modelType, error.err, correlationId))
             )
           );
         })
@@ -277,11 +282,11 @@ export class EntityOperators {
   deleteMany<TModel>() {
     return (source: Observable<DeleteMany<TModel>>) =>
       source.pipe(
-        mergeMap(({ info, entities, criteria }) => {
+        mergeMap(({ info, entities, criteria, correlationId }) => {
           return this.entityService.deleteMany<TModel>(info, entities, criteria).pipe(
-            map((ref: IEntityRef<TModel[]>) => new DeleteManySuccess<TModel>(ref.info.modelType, ref.entity)),
+            map((ref: IEntityRef<TModel[]>) => new DeleteManySuccess<TModel>(ref.info.modelType, ref.entity, correlationId)),
             catchError((error: IEntityError<TModel>) =>
-              handleError(error, new DeleteManyFailure<TModel>(error.info.modelType, error.err))
+              handleError(error, new DeleteManyFailure<TModel>(error.info.modelType, error.err, correlationId))
             )
           );
         })
@@ -290,66 +295,66 @@ export class EntityOperators {
 
   select<TModel>() {
     return (source: Observable<Select<TModel>>) =>
-      source.pipe(map(({ info, entity }) => new Selected<TModel>(info.modelType, entity)));
+      source.pipe(map(({ info, entity, correlationId }) => new Selected<TModel>(info.modelType, entity, correlationId)));
   }
 
   selectByKey<TModel>() {
     return (source: Observable<SelectByKey<TModel>>) =>
-      source.pipe(map(({ info, entityKey }) => new Selected<TModel>(info.modelType, entityKey)));
+      source.pipe(map(({ info, entityKey, correlationId }) => new Selected<TModel>(info.modelType, entityKey, correlationId)));
   }
 
   selectMany<TModel>() {
     return (source: Observable<SelectMany<TModel>>) =>
-      source.pipe(map(({ info, entities }) => new SelectedMany<TModel>(info.modelType, entities)));
+      source.pipe(map(({ info, entities, correlationId }) => new SelectedMany<TModel>(info.modelType, entities, correlationId)));
   }
 
   selectMore<TModel>() {
     return (source: Observable<SelectMore<TModel>>) =>
-      source.pipe(map(({ info, entities }) => new SelectedMore<TModel>(info.modelType, entities)));
+      source.pipe(map(({ info, entities, correlationId }) => new SelectedMore<TModel>(info.modelType, entities, correlationId)));
   }
 
   selectManyByKeys<TModel>() {
     return (source: Observable<SelectManyByKeys<TModel>>) =>
-      source.pipe(map(({ info, entitiesKeys }) => new SelectedMany<TModel>(info.modelType, entitiesKeys)));
+      source.pipe(map(({ info, entitiesKeys, correlationId }) => new SelectedMany<TModel>(info.modelType, entitiesKeys, correlationId)));
   }
 
   selectMoreByKeys<TModel>() {
     return (source: Observable<SelectMoreByKeys<TModel>>) =>
-      source.pipe(map(({ info, entitiesKeys }) => new SelectedMore<TModel>(info.modelType, entitiesKeys)));
+      source.pipe(map(({ info, entitiesKeys, correlationId }) => new SelectedMore<TModel>(info.modelType, entitiesKeys, correlationId)));
   }
 
   deselect<TModel>() {
     return (source: Observable<Deselect<TModel>>) =>
-      source.pipe(map(({ info }) => new Deselected<TModel>(info.modelType)));
+      source.pipe(map(({ info, correlationId }) => new Deselected<TModel>(info.modelType, correlationId)));
   }
 
   deselectMany<TModel>() {
     return (source: Observable<DeselectMany<TModel>>) =>
-      source.pipe(map(({ info, entities }) => new DeselectedMany<TModel>(info.modelType, entities)));
+      source.pipe(map(({ info, entities, correlationId }) => new DeselectedMany<TModel>(info.modelType, entities, correlationId)));
   }
 
   deselectManyByKeys<TModel>() {
     return (source: Observable<DeselectManyByKeys<TModel>>) =>
-      source.pipe(map(({ info, entitiesKeys }) => new DeselectedMany<TModel>(info.modelType, entitiesKeys)));
+      source.pipe(map(({ info, entitiesKeys, correlationId }) => new DeselectedMany<TModel>(info.modelType, entitiesKeys, correlationId)));
   }
 
   deselectAll<TModel>() {
     return (source: Observable<DeselectAll<TModel>>) =>
-      source.pipe(map(({ info }) => new DeselectedMany<TModel>(info.modelType, null)));
+      source.pipe(map(({ info, correlationId }) => new DeselectedMany<TModel>(info.modelType, null, correlationId)));
   }
 
   edit<TModel>() {
     return (source: Observable<Edit<TModel>>) =>
-      source.pipe(map(({ info, entity }) => new Edited<TModel>(info.modelType, entity)));
+      source.pipe(map(({ info, entity, correlationId }) => new Edited<TModel>(info.modelType, entity, correlationId)));
   }
 
   change<TModel>() {
     return (source: Observable<Change<TModel>>) =>
-      source.pipe(map(({ info, entity }) => new Changed<TModel>(info.modelType, entity)));
+      source.pipe(map(({ info, entity, correlationId }) => new Changed<TModel>(info.modelType, entity, correlationId)));
   }
 
   endEdit<TModel>() {
     return (source: Observable<EndEdit<TModel>>) =>
-      source.pipe(map(({ info }) => new EditEnded<TModel>(info.modelType)));
+      source.pipe(map(({ info, correlationId }) => new EditEnded<TModel>(info.modelType, correlationId)));
   }
 }